fix ListView sort not adhering to new GraphQL
