--- packages/core/src/___internal-do-not-use-will-break-in-patch/admin-ui/pages/ListPage/useSort.tsx
@@ -1,19 +1,0 @@
-import { useRouter } from 'next/router'
-
-import type { ListMeta } from '../../../../types'
-
-export function useSort(list: ListMeta) {
-  const { query } = useRouter()
-  const sortByFromUrl = typeof query.sortBy === 'string' ? query.sortBy : null
-  if (sortByFromUrl === '') return null
-  if (!sortByFromUrl) return list.initialSort
-  const fieldKey = sortByFromUrl.startsWith('-') ? sortByFromUrl.slice(1) : sortByFromUrl
-  const direction = sortByFromUrl.startsWith('-') ? ('DESC' as const) : ('ASC' as const)
-  const field = list.fields[fieldKey]
-  if (!field) return null
-  if (!field.isOrderable) return null
-  return {
-    field: fieldKey,
-    direction,
-  }
-}
